Added a function for ignoring network modules
Since currently the "get_layer_output_shapes" function will throw an
exception when encountering a module that is not present in the
"layer_type_to_function_map", there needs to be a way to ignore certain
module types, who don't affect the data shape. This can now be achieved
by adding such a module type to the dictionary with "__ignore_module"
function reference as value.